refactor solution to own file
